Add function to convert page count to bytes on Linux & Android
This commit adds a function that determines the page size on Linux and Android systems and uses this to convert virtual memory size and resident set size from pages to bytes. This was necessary because previously the reported values were in pages, which could lead to confusion or incorrect information when being used elsewhere in the code. By doing the conversion in this function, we ensure that the values are always correctly represented in bytes.